pdf_etl_pipeline/requirements.txt
# PDF ETL Pipeline Dependencies

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
tenacity>=8.0.0
pydantic>=2.0.0
orjson>=3.9.0
pandas>=2.0.0

# PDF processing and layout detection
layoutparser>=0.3.4
pdf2image>=3.1.0
Pillow>=10.0.0

# Machine learning backends for layoutparser
torch>=1.13.0
torchvision>=0.14.0
detectron2>=0.6.0

# Optional: for development
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0